Separate cluster controller related tables
Summary:
In this PR we take away all cluster and metadata related tables
away from the `restate-datafusion` package
