+++ crates/admin/Cargo.toml
@@ -17,6 +17,7 @@ replicated-loglet = ["restate-bifrost/replicated-loglet"]
 serve-web-ui = ["restate-web-ui", "mime_guess"]
 storage-query = []
 metadata-api = []
+table_docs = ["restate-datafusion/table_docs"]
 
 [dependencies]
 workspace-hack = { version = "0.1", path = "../../workspace-hack" }
