+++ crates/admin/src/cluster_controller/cluster_datafusion/partition/row.rs
@@ -8,10 +8,10 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use crate::table_util::format_using;
-
 use super::schema::PartitionBuilder;
 use itertools::{Itertools, Position};
+
+use restate_datafusion::table_util::format_using;
 use restate_types::{Version, cluster::cluster_state::RunMode, partition_table::Partition};
 
 #[inline]
