+++ crates/admin/src/cluster_controller/cluster_datafusion/partition_state/row.rs
@@ -8,12 +8,12 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use crate::table_util::format_using;
-
-use super::schema::PartitionStateBuilder;
+use restate_datafusion::table_util::format_using;
 use restate_types::{
     GenerationalNodeId, cluster::cluster_state::PartitionProcessorStatus, identifiers::PartitionId,
 };
+
+use super::schema::PartitionStateBuilder;
 
 #[inline]
 pub(crate) fn append_partition_row(
