+++ crates/admin/src/cluster_controller/mod.rs
@@ -8,6 +8,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+mod cluster_datafusion;
 pub mod cluster_state_refresher;
 pub mod grpc_svc_handler;
 mod logs_controller;
