+++ crates/admin/src/cluster_controller/service.rs
@@ -17,7 +17,7 @@ use anyhow::{Context, anyhow};
 use codederror::CodedError;
 use futures::never::Never;
 use restate_datafusion::BuildError;
-use restate_datafusion::context::{ClusterTables, QueryContext};
+use restate_datafusion::context::QueryContext;
 use restate_types::replication::{NodeSet, ReplicationProperty};
 use tokio::sync::{mpsc, oneshot};
 use tokio::time;
@@ -57,6 +57,7 @@ use restate_types::protobuf::common::AdminStatus;
 use restate_types::{GenerationalNodeId, NodeId, Version};
 
 use self::state::ClusterControllerState;
+use super::cluster_datafusion::ClusterTables;
 use super::cluster_state_refresher::ClusterStateRefresher;
 use super::grpc_svc_handler::ClusterCtrlSvcHandler;
 use super::protobuf::cluster_ctrl_svc_server::ClusterCtrlSvcServer;
