+++ crates/storage-query-datafusion/src/lib.rs
@@ -20,11 +20,6 @@ mod invocation_state;
 mod invocation_status;
 mod journal;
 mod keyed_service_status;
-mod log;
-mod node;
-mod node_state;
-mod partition;
-mod partition_state;
 pub mod partition_store_scanner;
 mod physical_optimizer;
 pub mod prelude;
